Change Component to release namespace of `render`
