+++ src/fuco/hook.ts
@@ -1,5 +1,4 @@
-import { Component, AttributeConverter } from "./component";
-import { HasCSSSymbol } from "./css";
+import { Component } from "./component";
 
 export interface Hooks<T> {
   _values: T[];
@@ -16,14 +15,11 @@ export type EffectFn = () => void | Cleanup;
 export type Cleanup = () => void;
 
 let currentCursor: number;
-let currentComponent: FucoComponent;
+let currentComponent: Component;
 
-export interface FucoComponent {
-  _hooks: Hooks<unknown>;
-  _attr<T>(name: string, converter?: AttributeConverter<T>): T | string | null;
-  _observeAttr(name: string, callback: () => void): void;
-  _dispatch<T>(name: string, init: CustomEventInit<T>): void;
-  _adoptStyle(css: HasCSSSymbol): void;
+export function __scope__(c: Component) {
+  currentComponent = c;
+  currentCursor = 0;
 }
 
 export function defaultHooks(): Hooks<unknown> {
@@ -36,23 +32,17 @@ export function defaultHooks(): Hooks<unknown> {
   };
 }
 
-export function __setCurrent__(c: FucoComponent) {
-  currentComponent = c;
-  currentCursor = 0;
-}
-
 export function hooks<T>(config: {
   _onmount?: (h: Hooks<T>, c: Component, i: number) => T;
   _onupdate?: (h: Hooks<T>, c: Component, i: number) => void;
 }): T {
   const h = currentComponent._hooks as Hooks<T>;
-  const c = currentComponent as Component;
   const index = currentCursor++;
   if (h._values.length <= index && config._onmount) {
-    h._values[index] = config._onmount(h, c, index);
+    h._values[index] = config._onmount(h, currentComponent, index);
   }
   if (config._onupdate) {
-    config._onupdate(h, c, index);
+    config._onupdate(h, currentComponent, index);
   }
   return h._values[index];
 }
