+++ src/fuco/index.ts
@@ -1,14 +1,17 @@
 export { html, render } from "../html";
 export { Component, AttributeConverter } from "./component";
-export {
-  defineElement,
-  FunctionalComponent,
-  __FucoRegistry__
-} from "./define-element";
+export { defineElement, FunctionalComponent, __def__ } from "./define-element";
 export { css, unsafeCSS } from "./css";
 export { Context } from "./context";
 export { createContext } from "./create-context";
-export { defaultHooks, Hooks, FucoComponent, __setCurrent__ } from "./hook";
+export {
+  __scope__,
+  defaultHooks,
+  Hooks,
+  Deps,
+  EffectFn,
+  Cleanup
+} from "./hook";
 export {
   useAttribute,
   useProperty,
